requirements.txt
discord.py==2.3.2
motor==3.3.2
pymongo==4.6.1
aiohttp==3.9.1
youtube-dl==2021.12.17
yt-dlp==2023.12.30
openai==1.6.1
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
asyncio==3.4.3
datetime
json5==0.9.14
beautifulsoup4==4.12.2
lxml==4.9.4
matplotlib==3.8.2
numpy==1.26.2
wavelink==3.4.1
spotipy==2.22.1
googletrans==4.0.0
profanity-check==1.0.3
better-profanity==0.7.0
langdetect==1.0.9
psutil==5.9.6
humanize==4.9.0
